pyproject.toml
[tool.poetry]
name = "MultiGit"
version = "0.1.0"
description = ""
authors = ["SkanderBM <skander.bmahmoud@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyGithub = "^1.55"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
